feat: unify ordered trie encoders
