+++ crates/trie/common/src/ordered_root.rs
@@ -1,257 +1,141 @@
-//! Incremental ordered trie root computation.
+//! Incremental ordered trie root computation for append-only streams.
 //!
-//! This module provides builders for computing ordered trie roots incrementally as items
-//! arrive, rather than requiring all items upfront. This is useful for receipt root
-//! calculation during block execution, where we know the total count but receive receipts
-//! one by one as transactions are executed.
+//! Ethereum block transaction and receipt roots are ordered trie roots keyed by the RLP encoding of
+//! each item index. [`OrderedTrieRootEncodedBuilder`] accepts items in their final contiguous
+//! order, without requiring the final item count until the stream ends.
+//!
+//! # Why this works
+//!
+//! RLP-encoded integer keys sort differently around small indices:
+//!
+//! - indices `1..=127` encode as single bytes `0x01..=0x7f`;
+//! - index `0` encodes as `0x80`, so it sorts after `1..=127`;
+//! - indices `>= 128` encode as longer byte strings and sort after `0`.
+//!
+//! This means the final insertion order for a dense list is:
+//!
+//! ```text
+//! len = 0:      []
+//! len = 1:      [0]
+//! 2..=128:     [1, 2, ..., len - 1, 0]
+//! >=129:       [1, 2, ..., 127, 0, 128, 129, ...]
+//! ```
+//!
+//! While the stream is open, the only ambiguous leaf is index `0`: for short lists it is inserted
+//! when the stream ends, and for lists with at least 129 items it can be inserted as soon as index
+//! `128` arrives. All other leaves can be added as they arrive in append-only order.
+//!
+//! # Example
+//!
+//! ```
+//! use reth_trie_common::ordered_root::OrderedTrieRootEncodedBuilder;
+//!
+//! let mut builder = OrderedTrieRootEncodedBuilder::new();
+//! builder.push_next(b"encoded_item_0");
+//! builder.push_next(b"encoded_item_1");
+//!
+//! let root = builder.finalize();
+//! ```
 
 use crate::{HashBuilder, Nibbles, EMPTY_ROOT_HASH};
 use alloc::vec::Vec;
 use alloy_primitives::B256;
-use alloy_trie::root::adjust_index_for_rlp;
-use core::fmt;
 
-/// Error returned when using [`OrderedTrieRootEncodedBuilder`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum OrderedRootError {
-    /// Called `finalize()` before all items were pushed.
-    Incomplete {
-        /// The expected number of items.
-        expected: usize,
-        /// The number of items received.
-        received: usize,
-    },
-    /// Index is out of bounds.
-    IndexOutOfBounds {
-        /// The index that was provided.
-        index: usize,
-        /// The expected length.
-        len: usize,
-    },
-    /// Item at this index was already pushed.
-    DuplicateIndex {
-        /// The duplicate index.
-        index: usize,
-    },
-}
+/// First index whose RLP-encoded key sorts after index 0.
+///
+/// Indices `1..=0x7f` encode as their own single byte, so they sort before the RLP encoding of
+/// index 0 (`0x80`). Index `0x80` and larger use long-form RLP integer encoding and sort after
+/// index 0, so index 0 must be flushed before inserting this index.
+const ZERO_KEY_FLUSH_INDEX: usize = 0x80;
 
-impl OrderedRootError {
-    /// Returns `true` if the error is [`OrderedRootError::Incomplete`].
-    #[inline]
-    pub const fn is_incomplete(&self) -> bool {
-        matches!(self, Self::Incomplete { .. })
-    }
-
-    /// Returns `true` if the error is [`OrderedRootError::IndexOutOfBounds`].
-    #[inline]
-    pub const fn is_index_out_of_bounds(&self) -> bool {
-        matches!(self, Self::IndexOutOfBounds { .. })
-    }
-
-    /// Returns `true` if the error is [`OrderedRootError::DuplicateIndex`].
-    #[inline]
-    pub const fn is_duplicate_index(&self) -> bool {
-        matches!(self, Self::DuplicateIndex { .. })
-    }
-
-    /// Returns the index associated with the error, if any.
-    #[inline]
-    pub const fn index(&self) -> Option<usize> {
-        match self {
-            Self::Incomplete { .. } => None,
-            Self::IndexOutOfBounds { index, .. } | Self::DuplicateIndex { index } => Some(*index),
-        }
-    }
-}
-
-impl fmt::Display for OrderedRootError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Incomplete { expected, received } => {
-                write!(f, "incomplete: expected {expected} items, received {received}")
-            }
-            Self::IndexOutOfBounds { index, len } => {
-                write!(f, "index {index} out of bounds for length {len}")
-            }
-            Self::DuplicateIndex { index } => {
-                write!(f, "duplicate item at index {index}")
-            }
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for OrderedRootError {}
-
-/// A builder for computing ordered trie roots incrementally from pre-encoded items.
+/// A builder for computing ordered trie roots from an append-only stream of pre-encoded items.
 ///
-/// This builder allows pushing items one by one as they become available
-/// (e.g., receipts after each transaction execution), rather than requiring
-/// all items upfront.
+/// This type does not require the total item count up front. It relies on a strict input contract
+/// instead: every pushed item must be the next contiguous item in the final list.
 ///
-/// # Use Case
-///
-/// When executing a block, the receipt root must be computed from all transaction
-/// receipts. With the standard `ordered_trie_root`, you must wait until all
-/// transactions are executed before computing the root. This builder enables
-/// **incremental computation** - you can start building the trie as soon as
-/// receipts become available, potentially in parallel with continued execution.
-///
-/// The builder requires knowing the total item count upfront (the number of
-/// transactions in the block), but items can be pushed in any order by index.
-///
-/// # How It Works
-///
-/// Items can be pushed in any order by specifying their index. The builder
-/// internally buffers items and flushes them to the underlying [`HashBuilder`]
-/// in the correct order for RLP key encoding (as determined by [`adjust_index_for_rlp`]).
-///
-/// # Memory
-///
-/// Each pushed item is stored in an internal buffer until it can be flushed.
-/// In the worst case (e.g., pushing index 0 last), all items except one will
-/// be buffered. For receipt roots, index 0 is typically flushed late due to
-/// RLP key ordering, so expect to buffer most items until near the end.
-///
-/// # Example
-///
-/// ```
-/// use reth_trie_common::ordered_root::OrderedTrieRootEncodedBuilder;
-///
-/// // Create a builder for 2 pre-encoded items
-/// let mut builder = OrderedTrieRootEncodedBuilder::new(2);
-///
-/// // Push pre-encoded items as they arrive (can be out of order)
-/// builder.push(1, b"encoded_item_1").unwrap();
-/// builder.push(0, b"encoded_item_0").unwrap();
-///
-/// // Finalize to get the root hash
-/// let root = builder.finalize().unwrap();
-/// ```
-#[derive(Debug)]
+/// This builder is intended for transaction and receipt root computation while a block is still
+/// being built. It buffers only index `0`, then streams all other leaves directly into the
+/// [`HashBuilder`] in the same order as `alloy_trie::root::ordered_trie_root_encoded`.
+#[derive(Debug, Default)]
 pub struct OrderedTrieRootEncodedBuilder {
-    /// Total expected number of items.
+    /// Number of items pushed so far. This is also the next append-only index.
     len: usize,
-    /// Number of items received so far.
-    received: usize,
-    /// Next insertion loop counter (determines which adjusted index to flush next).
-    next_insert_i: usize,
-    /// Buffer for pending items, indexed by execution index.
-    pending: Vec<Option<Vec<u8>>>,
+    /// Index 0 is the only item whose final insertion position depends on whether more items
+    /// arrive.
+    zero: Option<Vec<u8>>,
     /// The underlying hash builder.
     hb: HashBuilder,
 }
 
 impl OrderedTrieRootEncodedBuilder {
-    /// Creates a new builder for `len` pre-encoded items.
-    pub fn new(len: usize) -> Self {
-        Self {
-            len,
-            received: 0,
-            next_insert_i: 0,
-            pending: alloc::vec![None; len],
-            hb: HashBuilder::default(),
-        }
+    /// Creates an empty builder.
+    pub fn new() -> Self {
+        Self::default()
     }
 
-    /// Pushes a pre-encoded item at the given index to the builder.
+    /// Pushes the next pre-encoded item.
     ///
-    /// Items can be pushed in any order. The builder will automatically
-    /// flush items to the underlying [`HashBuilder`] when they become
-    /// available in the correct order.
-    ///
-    /// # Errors
-    ///
-    /// - [`OrderedRootError::IndexOutOfBounds`] if `index >= len`
-    /// - [`OrderedRootError::DuplicateIndex`] if an item was already pushed at this index
+    /// Items must be pushed in their final contiguous order.
     #[inline]
-    pub fn push(&mut self, index: usize, bytes: &[u8]) -> Result<(), OrderedRootError> {
-        if index >= self.len {
-            return Err(OrderedRootError::IndexOutOfBounds { index, len: self.len });
+    pub fn push_next(&mut self, bytes: &[u8]) {
+        let index = self.len;
+        self.len += 1;
+
+        match index {
+            0 => {
+                self.zero = Some(bytes.to_vec());
+            }
+            1..=0x7f => {
+                self.add_leaf(index, bytes);
+            }
+            ZERO_KEY_FLUSH_INDEX => {
+                self.flush_zero();
+                self.add_leaf(index, bytes);
+            }
+            _ => {
+                debug_assert!(self.zero.is_none(), "index 0 must be flushed before indices > 128");
+                self.add_leaf(index, bytes);
+            }
         }
-
-        if self.pending[index].is_some() {
-            return Err(OrderedRootError::DuplicateIndex { index });
-        }
-
-        self.push_unchecked(index, bytes);
-        Ok(())
-    }
-
-    /// Pushes a pre-encoded item at the given index without bounds or duplicate checking.
-    ///
-    /// This is a performance-critical method for callers that can guarantee:
-    /// - `index < len`
-    /// - No item has been pushed at this index before
-    ///
-    /// # Panics
-    ///
-    /// Panics in debug mode if `index >= len`.
-    #[inline]
-    pub fn push_unchecked(&mut self, index: usize, bytes: &[u8]) {
-        debug_assert!(index < self.len, "index {index} out of bounds for length {}", self.len);
-        debug_assert!(self.pending[index].is_none(), "duplicate item at index {index}");
-
-        self.pending[index] = Some(bytes.to_vec());
-        self.received += 1;
-
-        self.flush();
-    }
-
-    /// Attempts to flush pending items to the hash builder.
-    fn flush(&mut self) {
-        while self.next_insert_i < self.len {
-            let exec_index_needed = adjust_index_for_rlp(self.next_insert_i, self.len);
-
-            let Some(value) = self.pending[exec_index_needed].take() else {
-                break;
-            };
-
-            let index_buffer = alloy_rlp::encode_fixed_size(&exec_index_needed);
-            self.hb.add_leaf(Nibbles::unpack(&index_buffer), &value);
-
-            self.next_insert_i += 1;
-        }
-    }
-
-    /// Returns `true` if all items have been pushed.
-    #[inline]
-    pub const fn is_complete(&self) -> bool {
-        self.received == self.len
     }
 
     /// Returns the number of items pushed so far.
     #[inline]
     pub const fn pushed_count(&self) -> usize {
-        self.received
+        self.len
     }
 
-    /// Returns the expected total number of items.
+    /// Returns `true` if no items have been pushed.
     #[inline]
-    pub const fn expected_count(&self) -> usize {
-        self.len
+    pub const fn is_empty(&self) -> bool {
+        self.len == 0
     }
 
     /// Finalizes the builder and returns the trie root.
     ///
-    /// # Errors
-    ///
-    /// Returns [`OrderedRootError::Incomplete`] if not all items have been pushed.
-    pub fn finalize(mut self) -> Result<B256, OrderedRootError> {
+    /// This method is the end-of-stream signal. If fewer than 129 items were pushed, index `0` is
+    /// inserted here because it is the final leaf in RLP key order for that short list.
+    pub fn finalize(mut self) -> B256 {
         if self.len == 0 {
-            return Ok(EMPTY_ROOT_HASH);
+            return EMPTY_ROOT_HASH;
         }
 
-        if self.received != self.len {
-            return Err(OrderedRootError::Incomplete {
-                expected: self.len,
-                received: self.received,
-            });
+        self.flush_zero();
+        self.hb.root()
+    }
+
+    fn flush_zero(&mut self) {
+        if self.zero.is_none() {
+            return;
         }
 
-        debug_assert_eq!(self.next_insert_i, self.len, "not all items were flushed");
+        let zero = self.zero.take().expect("index 0 must be buffered before it is flushed");
+        self.add_leaf(0, &zero);
+    }
 
-        Ok(self.hb.root())
+    fn add_leaf(&mut self, index: usize, bytes: &[u8]) {
+        let index_buffer = alloy_rlp::encode_fixed_size(&index);
+        self.hb.add_leaf(Nibbles::unpack(&index_buffer), bytes);
     }
 }
 
@@ -259,6 +143,23 @@ impl OrderedTrieRootEncodedBuilder {
 mod tests {
     use super::*;
     use alloy_trie::root::ordered_trie_root_encoded;
+    use proptest::prelude::*;
+
+    fn item(index: usize) -> Vec<u8> {
+        format!("encoded_item_{index}").into_bytes()
+    }
+
+    fn items(len: usize) -> Vec<Vec<u8>> {
+        (0..len).map(item).collect()
+    }
+
+    fn root_with_push_next(items: &[Vec<u8>]) -> B256 {
+        let mut builder = OrderedTrieRootEncodedBuilder::new();
+        for item in items {
+            builder.push_next(item);
+        }
+        builder.finalize()
+    }
 
     #[test]
     fn test_ordered_encoded_builder_equivalence() {
@@ -268,13 +169,13 @@ mod tests {
 
             let expected = ordered_trie_root_encoded(&items);
 
-            let mut builder = OrderedTrieRootEncodedBuilder::new(len);
+            let mut builder = OrderedTrieRootEncodedBuilder::new();
 
-            for (i, item) in items.iter().enumerate() {
-                builder.push(i, item).unwrap();
+            for item in &items {
+                builder.push_next(item);
             }
 
-            let actual = builder.finalize().unwrap();
+            let actual = builder.finalize();
             assert_eq!(
                 expected, actual,
                 "mismatch for len={len}: expected {expected:?}, got {actual:?}"
@@ -283,72 +184,132 @@ mod tests {
     }
 
     #[test]
-    fn test_ordered_builder_out_of_order() {
-        for len in [2, 3, 5, 10, 50] {
-            let items: Vec<Vec<u8>> =
-                (0..len).map(|i| format!("item_{i}_data").into_bytes()).collect();
+    fn empty_builder_returns_empty_root() {
+        let builder = OrderedTrieRootEncodedBuilder::new();
+        assert!(builder.is_empty());
+        assert_eq!(builder.pushed_count(), 0);
+        assert_eq!(builder.finalize(), EMPTY_ROOT_HASH);
+    }
 
+    #[test]
+    fn test_ordered_builder_count_tracks_pushed_items() {
+        let mut builder = OrderedTrieRootEncodedBuilder::new();
+        assert!(builder.is_empty());
+
+        builder.push_next(b"item_0");
+        assert!(!builder.is_empty());
+        assert_eq!(builder.pushed_count(), 1);
+
+        builder.push_next(b"item_1");
+        assert_eq!(builder.pushed_count(), 2);
+    }
+
+    #[test]
+    fn matches_ordered_root_for_boundary_lengths() {
+        for len in [0, 1, 2, 3, 4, 126, 127, 128, 129, 130, 200, 255, 256, 512] {
+            let items = items(len);
             let expected = ordered_trie_root_encoded(&items);
 
-            // Push in reverse order
-            let mut builder = OrderedTrieRootEncodedBuilder::new(len);
-            for i in (0..len).rev() {
-                builder.push(i, &items[i]).unwrap();
-            }
-            let actual = builder.finalize().unwrap();
-            assert_eq!(expected, actual, "mismatch for reverse order len={len}");
-
-            // Push odds first, then evens
-            let mut builder = OrderedTrieRootEncodedBuilder::new(len);
-            for i in (1..len).step_by(2) {
-                builder.push(i, &items[i]).unwrap();
-            }
-            for i in (0..len).step_by(2) {
-                builder.push(i, &items[i]).unwrap();
-            }
-            let actual = builder.finalize().unwrap();
-            assert_eq!(expected, actual, "mismatch for odd/even order len={len}");
+            assert_eq!(root_with_push_next(&items), expected, "push_next mismatch for len={len}");
         }
     }
 
     #[test]
-    fn test_ordered_builder_empty() {
-        let builder = OrderedTrieRootEncodedBuilder::new(0);
-        assert!(builder.is_complete());
-        assert_eq!(builder.finalize().unwrap(), EMPTY_ROOT_HASH);
+    fn short_lists_flush_zero_on_finalize() {
+        for len in 1..=128 {
+            let items = items(len);
+            let expected = ordered_trie_root_encoded(&items);
+
+            let mut builder = OrderedTrieRootEncodedBuilder::new();
+            for item in &items {
+                builder.push_next(item);
+            }
+
+            assert_eq!(builder.pushed_count(), len);
+            assert_eq!(builder.finalize(), expected, "mismatch for short len={len}");
+        }
     }
 
     #[test]
-    fn test_ordered_builder_incomplete_error() {
-        let mut builder = OrderedTrieRootEncodedBuilder::new(3);
+    fn long_lists_flush_zero_when_index_128_arrives() {
+        let items = items(129);
+        let expected = ordered_trie_root_encoded(&items);
 
-        builder.push(0, b"item_0").unwrap();
-        builder.push(1, b"item_1").unwrap();
+        let mut builder = OrderedTrieRootEncodedBuilder::new();
+        for item in &items {
+            builder.push_next(item);
+        }
 
-        assert!(!builder.is_complete());
-        assert_eq!(
-            builder.finalize(),
-            Err(OrderedRootError::Incomplete { expected: 3, received: 2 })
-        );
+        assert_eq!(builder.pushed_count(), 129);
+        assert_eq!(builder.finalize(), expected);
     }
 
-    #[test]
-    fn test_ordered_builder_index_errors() {
-        let mut builder = OrderedTrieRootEncodedBuilder::new(2);
+    proptest! {
+        #[test]
+        fn arbitrary_encoded_items_match_ordered_root(items in proptest::collection::vec(
+            proptest::collection::vec(any::<u8>(), 0..128),
+            0..1024,
+        )) {
+            let expected = ordered_trie_root_encoded(&items);
 
-        assert_eq!(
-            builder.push(5, b"item"),
-            Err(OrderedRootError::IndexOutOfBounds { index: 5, len: 2 })
-        );
+            prop_assert_eq!(root_with_push_next(&items), expected);
+        }
+    }
 
-        builder.push(0, b"item_0").unwrap();
+    #[cfg(feature = "arbitrary")]
+    mod arbitrary_consensus_roots {
+        use super::*;
+        use alloy_consensus::{
+            proofs::{calculate_receipt_root, calculate_transaction_root},
+            EthereumReceipt, ReceiptWithBloom, Signed, TxLegacy,
+        };
+        use alloy_eips::eip2718::Encodable2718;
+        use alloy_primitives::Signature;
+        use proptest::test_runner::Config;
+        use proptest_arbitrary_interop::arb;
 
-        assert_eq!(
-            builder.push(0, b"item_0_dup"),
-            Err(OrderedRootError::DuplicateIndex { index: 0 })
-        );
+        fn open_ended_2718_root<T: Encodable2718>(items: &[T]) -> B256 {
+            let mut builder = OrderedTrieRootEncodedBuilder::new();
+            let mut buf = Vec::new();
 
-        builder.push(1, b"item_1").unwrap();
-        assert!(builder.is_complete());
+            for item in items {
+                buf.clear();
+                item.encode_2718(&mut buf);
+                builder.push_next(&buf);
+            }
+
+            builder.finalize()
+        }
+
+        proptest! {
+            #![proptest_config(Config::with_cases(32))]
+
+            #[test]
+            fn arbitrary_transactions_match_alloy_consensus_root(
+                transactions in proptest::collection::vec(
+                    (arb::<TxLegacy>(), arb::<Signature>())
+                        .prop_map(|(tx, signature)| Signed::new_unhashed(tx, signature)),
+                    0..1024,
+                ),
+            ) {
+                let expected = calculate_transaction_root(&transactions);
+                let actual = open_ended_2718_root(&transactions);
+
+                prop_assert_eq!(actual, expected);
+            }
+
+            #[test]
+            fn arbitrary_receipts_match_alloy_consensus_root(
+                receipts in proptest::collection::vec(
+                    arb::<ReceiptWithBloom<EthereumReceipt>>(),
+                    0..1024,
+                ),
+            ) {
+                let expected = calculate_receipt_root(&receipts);
+                let actual = open_ended_2718_root(&receipts);
+
+                prop_assert_eq!(actual, expected);
+            }
+        }
     }
 }
